bump version

+++ src/main/java/com/infraly/sourcequery/protocol/A2SProtocol.java
@@ -14,7 +14,7 @@ public final class A2SProtocol {
     public static final byte RESPONSE_RULES = 0x45;
     public static final byte RESPONSE_CHALLENGE = 0x41;
 
-    public static final byte GAMEPORT_FLAG = 0x80;
+    public static final byte GAMEPORT_FLAG = (byte) 0x80;
 
     public static final byte SERVER_TYPE_DEDICATED = 'd';
